+++ pynonymizer/__main__.py
@@ -81,7 +81,7 @@ def create_parser():
                         help="Choose a step to stop at (inclusive). [$PYNONYMIZER_STOP_AT]")
 
     parser.add_argument("--seed-rows",
-                        default=os.getenv("PYNONYMIZER_SEED_ROWS"),
+                        default=os.getenv("PYNONYMIZER_SEED_ROWS"), type=int,
                         help="Specify a number of rows to populate the fake data table used during anonymization.  [$PYNONYMIZER_SEED_ROWS]")
 
     parser.add_argument("--mssql-backup-compression",
@@ -94,6 +94,10 @@ def create_parser():
     parser.add_argument("--verbose", action="store_true", default=os.getenv("PYNONYMISER_VERBOSE") or False,
                         help="Increases the verbosity of the logging feature, to help when troubleshooting issues. [$PYNONYMIZER_VERBOSE]"
                         )
+
+    parser.add_argument("--dry-run",
+                        default=os.getenv("PYNONYMIZER_DRY_RUN") or False, action="store_true",
+                        help="Instruct pynonymizer to skip all process steps. Useful for testing safely.  [$PYNONYMIZER_DRY_RUN]")
 
     return parser
 
@@ -133,7 +137,9 @@ def main(rawArgs=None):
             skip_steps=args.skip_steps,
             stop_at_step=args.stop_at_step,
             seed_rows=args.seed_rows,
-            mssql_backup_compression=args.mssql_backup_compression
+            mssql_backup_compression=args.mssql_backup_compression,
+            dry_run=args.dry_run,
+            verbose=args.verbose
         )
     except ModuleNotFoundError as error:
         if error.name == "pyodbc" and args.db_type == "mssql":