+++ pynonymizer/database/postgres/execution.py
@@ -7,7 +7,7 @@ Seperate everything that touches actual query exec into its own module
 """
 
 class PSqlDumpRunner:
-    def __init__(self, db_host, db_user, db_pass, db_name, db_port=5432):
+    def __init__(self, db_host, db_user, db_pass, db_name, db_port="5432"):
         self.db_host = db_host
         self.db_user = db_user
         self.db_pass = db_pass
@@ -34,7 +34,7 @@ class PSqlDumpRunner:
 
 
 class PSqlCmdRunner:
-    def __init__(self, db_host, db_user, db_pass, db_name, db_port=5432):
+    def __init__(self, db_host, db_user, db_pass, db_name, db_port="5432"):
         self.db_host = db_host
         self.db_user = db_user
         self.db_pass = db_pass